+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a CHANGELOG](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-Upcoming v0.4.7: Log source is tus code
+Upcoming v0.5.0: Log source is tus code
+- move Responsorial to TusClient namespace
 - move HttpService to lib/, tusc/ contains tus code.
 - tests use std log (log/tusc.log) vs. log/tusc_test.log